Add connection directly + UI improvements
For chrome specific tasks (eg. add breakpoint) fetch a connection
directly. This allows us to prompt for connection then execute
breakpoint tasks (this isn't possible with regular get-client! behaviour
as we need to check for the script that we expect to have been parsed
already - this fails and no breakpoint call is executed).

Significant UI changes / improvements